1-level working + some progress in nesting lists

+++ inspect.sh
@@ -1,10 +1,20 @@
 #!/usr/bin/bash
 TEMPFILE=/tmp/content.xml
-# -- clenup
+DOCUMENT=$1
+
+if [[ -z "$DOCUMENT" ]];
+then
+	echo "You need to specify a document to inspect"
+	exit 0
+fi
+
+# -- cleanup
 [[ -f $TEMPFILE ]] && rm $TEMPFILE
+
 # extract one file from archive, lint it and save to temporary file. Then edit it with VIM
-# 7z x -so /tmp/Test_output.odt content.xml | xmllint --format - > $TEMPFILE && nvim -Rn $TEMPFILE
-7z x -so /tmp/test.odt content.xml | 
+7z x -so "$DOCUMENT" content.xml | 
 	xmllint --format - |
 	sgrep -o "%r\n" '"<office:body>".."</office:body>"' |
 	pygmentize
+
+exit 0
